requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==45.0.5
Django==4.2.7
django-crispy-forms==2.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
fsspec==2025.7.0
huggingface-hub==0.33.4
idna==3.10
iopath==0.1.10
layoutparser==0.3.4
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pdfrw==0.4
Pillow==10.1.0
portalocker==3.2.0
pycparser==2.22
PyJWT==2.10.1
PyMuPDF==1.26.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pywin32==311
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.0.7
requests==2.32.4
safetensors==0.5.3
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.53.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
